+++ Changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 ### Fixed
+ - Fixing yearRange option for angular 1 renderer which was switched to yearRows and yearColumns.
 
 ### Removed