+++ Changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 ### Fixed
+ - Fixed min/max validation issues with the datetime component
 
 ### Removed
  - Removed the m/d/y selection modes for the date time component, because they are only supported in the angular 1