+++ Changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - Added the layout components from resources to display in the Existing Resource Fields select list
 
 ### Changed
+ - Upgrade ng-formio to 2.15.8
 
 ### Fixed
  - Fixing yearRange option for angular 1 renderer which was switched to yearRows and yearColumns.